backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
httpx==0.28.1
feedparser==6.0.11
lark-oapi==1.4.5
chromadb==0.6.3
apscheduler==3.10.4
python-dotenv==1.0.1
